python-project/pyproject.toml
[project]
name = "mortar"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "mk",
    "pillow",
    "PyOpenGL",
    "tomlkit"
]

[project.optional-dependencies]
dev = [
    "coverage",
    "flake8",
    "mypy",
    "nox",
    "pytest",
    "types-pillow"
]

[project.scripts]
mortess = "mortar.tesseract:main"
mortvids = "mortar.video:main"

[tool.coverage.run]
source = [ "mortar" ]

[tool.mypy]
explicit_package_bases = true
packages = [ "mortar", "test" ]
strict = true

[tool.pytest.ini_options]
testpaths = [ "test" ]
